feat: Create service for the domain
